% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currency_codes.R
\name{get_currency_codes}
\alias{get_currency_codes}
\title{get currency codes}
\usage{
get_currency_codes()
}
\value{
data.frame with currency codes
}
\description{
gets currency codes from nbu website
}

