% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_exchange_rate.R
\name{get_current_exchange_rate}
\alias{get_current_exchange_rate}
\title{get exchange rate for all currencies for current date}
\usage{
get_current_exchange_rate()
}
\value{
dataframe with exchange rate for all available currencies for current date
}
\description{
get exchange rate for all currencies for current date
}

