% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rate_date.R
\name{get_exchange_rate_date}
\alias{get_exchange_rate_date}
\title{get exchange rate for specific date (in YYYYMMDD format)}
\usage{
get_exchange_rate_date(date)
}
\arguments{
\item{date}{- date in yyyymmdd format}
}
\value{
dataframe with currency exchange rate
}
\description{
get exchange rate for specific date (in YYYYMMDD format)
}

