% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rate_date_currency.R
\name{get_exchange_rate_date_currency}
\alias{get_exchange_rate_date_currency}
\title{get exchange rate for specific date (in YYYYMMDD format) and currency}
\usage{
get_exchange_rate_date_currency(date, currency)
}
\arguments{
\item{date}{- date in yyyymmdd format}

\item{currency}{- currency code}
}
\value{
data.frame with currency exchange rate
}
\description{
get exchange rate for specific date (in YYYYMMDD format) and currency
}

