% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rate_date_range.R
\name{get_exchange_rate_date_range}
\alias{get_exchange_rate_date_range}
\title{get exchange rate for all currencies and given date range}
\usage{
get_exchange_rate_date_range(start, end)
}
\arguments{
\item{start}{- date in yyyymmdd format}

\item{end}{- date in yyyymmdd format}
}
\value{
data.frame with exchange rates
}
\description{
get exchange rate for all currencies and given date range
}

