% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_rate_date_range_currency.R
\name{get_exchange_rate_date_range_currency}
\alias{get_exchange_rate_date_range_currency}
\title{get exchange rate for specific currency and given date range}
\usage{
get_exchange_rate_date_range_currency(start, end, currency)
}
\arguments{
\item{start}{- date in yyyymmdd format}

\item{end}{- date in yyyymmdd format}

\item{currency}{- currency code}
}
\value{
data.frame with exchange rates
}
\description{
get exchange rate for specific currency and given date range
}

